# core/connection: remove unnecessary defers