@skriptble 2018-06-15
Hi @odeke-em,

Thanks for the contribution! This looks fine to me, I'll merge it.

--Kris